{"schema":"decodex/commit/1","summary":"Modularize capture-frame model types","authority":"manual"}
